+++ R/nmr_normalize.R
@@ -67,7 +67,7 @@ norm_pqn <- function(spectra) {
 #' nmr_dataset <- nmr_dataset_load(system.file("extdata", "nmr_dataset.rds", package = "AlpsNMR"))
 #' nmr_dataset <- nmr_normalize(nmr_dataset, method = "area")
 #' norm_dataset <- nmr_normalize(nmr_dataset)
-#' print(norm_dataset$plot)
+#' norm_dataset$plot
 nmr_normalize <- function(samples, 
                           method = c("area", "max", "value", "region", "pqn", "none"),
                           ...) {
@@ -115,7 +115,7 @@ nmr_normalize <- function(samples,
 #' nmr_dataset <- nmr_dataset_load(system.file("extdata", "nmr_dataset.rds", package = "AlpsNMR"))
 #' nmr_dataset <- nmr_normalize(nmr_dataset, method = "area")
 #' norm_extra_info <- nmr_normalize_extra_info(nmr_dataset)
-#' print(norm_extra_info$plot)
+#' norm_extra_info$plot
 nmr_normalize_extra_info <- function(samples) {
     attr(samples, "normalize_extra_info")
 }